style: move Sunburst into its own folder
