+++ src/pages/ReportPage/ReportPage.tsx
@@ -6,7 +6,7 @@ import {
   type Node,
   type Path,
 } from "../../utils/zon";
-import Sunburst from "./Sunburst.tsx";
+import Sunburst from "./Sunburst/Sunburst.tsx";
 import ReportList from "./ReportList/ReportList.tsx";
 import UploadButton from "../../components/UploadButton/UploadButton.tsx";
 import Breadcrumbs from "./Breadcrumbs/Breadcrumbs.tsx";
